Bonjour SDK for Windows v3.0
The Bonjour SDK for Windows includes the latest version of Bonjour
as well as header files, libraries, and sample code to assist
developers in creating Bonjour enabled applications on Windows.
The SDK has been updated with the Bonjour core that is bundled with
iTunes 10.3.1. This release will bring all existing Bonjour functionality
released in Mac OS X 10.7 into the Bonjour for Windows product.

https://developer.apple.com/download/all/?q=Bonjour%20SDK%20for%20Windows